backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
sentence-transformers==2.2.2
clip-by-openai==1.0
open-clip-torch==2.20.0
pytesseract==0.3.10
easyocr==1.7.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
faiss-cpu==1.7.4
pyyaml==6.0.1
aiofiles==23.2.1
python-dateutil==2.8.2
networkx==3.2.1
inflect==7.0.0
httpx==0.25.2
openai==1.3.5
pytest==7.4.3
pytest-asyncio==0.21.1
